Fine

+++ tasks/variables/car_price.py
@@ -30,7 +30,7 @@ def car_price(price: float) -> float:
     :return: Стоимость автомобиля с наценками
     """
 
-    result = float(price + price * 0.18 + price * 0.05 + 500 + 100)
+    result = price + price * 0.18 + price * 0.05 + 500 + 100
     return result
